Stoppa pagination om du sorterat på en kategori Staffindex - Edit fungerar inte, tagit bort . GetAllAsync() som hämtade alla filmer tas bort. Detta förhindrar att Index-sidan laddar långsamt när du har tusentals poster.
GetAllAsync(int pageNumber, int pageSize) och CountAsync() tas också bort.
Nu hanteras allt inom en enda metod GetPagedAsync() som returnerar:

Films: listan över filmer för den aktuella sidan

TotalCount: totalt antal filmer som matchar filtret (för att beräkna sidorna)

GetPagedAsync() accepterar:

search: sökning efter titel eller skådespelare

sortOrder: sortering

pageNumber och pageSize: paginering

+++ Repository/IStaffRepository.cs
@@ -5,7 +5,7 @@ namespace RetroVHSRental.Repository
     public interface IStaffRepository
     {
         Task<IEnumerable<Staff>> GetAllAsync();
-        Task<Staff> GetByIdAsync(int id);
+        Task<Staff> GetByIdAsync(byte id);
         Task AddAsync(Staff staff);
         Task RemoveAsync(Staff staff);
         Task UpdateAsync(Staff staff);
